refactor: wizard steps handling

+++ src/app/shared/modal/modal-host.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit, ViewChild, ComponentFactoryResolver } from '@angular/core';
 import { ContentHostDirective } from '../directives/content-host/content-host.directive';
-import { ModalService, ModalState } from './modal.service';
+import { ModalService } from './modal.service';
+import { ModalState, ModalModes } from "./modalState";
 import { ModalComponent } from './modal.component';
 import { DialogService } from '../dialog/dialog.service';
 
@@ -34,22 +35,25 @@ export class ModalHostComponent implements OnInit {
         state.handleSave(eventData);
         this.dismissModal(refs);
       });
-      refs.component.deleteClicked.subscribe(eventData => {
-        this.dialogService.confirm({
-          heading: state.dialog.heading,
-          message: state.dialog.message,
-          callback: affirmativeAnswer => {
-            if (affirmativeAnswer) {
-              state.handleDelete(eventData);
-              this.dismissModal(refs);
+      if (state.handleDelete) {
+        refs.component.deleteClicked.subscribe(eventData => {
+          this.dialogService.confirm({
+            heading: state.dialog.heading,
+            message: state.dialog.message,
+            callback: affirmativeAnswer => {
+              if (affirmativeAnswer) {
+                state.handleDelete(eventData);
+                this.dismissModal(refs);
+              }
             }
-          }
+          });
         });
-      });
+      }
       refs.component.content = {
         component: state.component,
         formData: state.formData,
         heading: state.heading,
+        showDelete: state.modalMode == ModalModes.update
       };
     };
   }
